requirements.txt
astroid==2.4.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
toml==0.10.2
twitter==1.18.0
urllib3==1.26.2
wrapt==1.12.1
